requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40
lightgbm>=4.0
scikit-learn>=1.3
mlflow>=2.8
streamlit>=1.30
plotly>=5.18
duckdb>=0.10
pyarrow>=14.0
pyyaml>=6.0
torch>=2.2

